Client QOL improvements
Hello. I was interested in using the library and noticed a few things I was originally going to address on a personal fork before deciding to make a PR; I wanted to allow the bot to connect to localhost PS instances for testing and to improve how  worked wrt. removing the temporary listeners to prevent an unbounded number of listeners needlessly lurking and running on the same events. I also made it so that listening for the challstr message is a one-off listener rather than adding it as a permanent listener, which seemed needless to me.